+++ handlers/handleWatchlistRequest.ts
@@ -9,12 +9,12 @@ import { getTraktRecentWatches } from "../services/trakt.ts";
 import { SEARCH_COUNT } from "../config/env.ts";
 import { updateRpdbPosters } from "../services/rpdb.ts";
 import { getProviderInfoFromState } from "../services/aiProvider.ts";
-import { pushBatchToQstash } from "../config/qstash.ts";
+import { ListTaskParams, pushBatchToQstash, pushListToQstash } from "../config/qstash.ts";
 import { createRedisKey } from "../services/tmdbHelpers/tmdbCommon.ts";
 import { isOldCacheStructure, convertOldToNewStructure } from "../services/tmdbHelpers/fixOldCache.ts";
 
 export const handleTraktWatchlistRequest = async (ctx: Context) => {
-  const { tmdbKey, googleKey, openAiKey, traktKey, rpdbKey, omdbKey, userId, type } = ctx.state;
+  const { tmdbKey, googleKey, openAiKey, traktKey, rpdbKey, omdbKey, userId, type, traktCreateList } = ctx.state;
   
   // Validate required parameters
   if (!traktKey || !type || !userId || !tmdbKey || (!googleKey && !openAiKey)) {
@@ -93,7 +93,7 @@ export const handleTraktWatchlistRequest = async (ctx: Context) => {
               backgroundUpdateBatch.push({ movieName, lang, type, tmdbKey, omdbKey, redisKey });
               tmdbData = convertOldToNewStructure(parsed, type);
             } else {
-              tmdbData = parsed;
+              tmdbData = parsed as Meta;
             }
             stats.fromCache++;
           } catch (err) {
@@ -118,7 +118,19 @@ export const handleTraktWatchlistRequest = async (ctx: Context) => {
     // Filter out any invalid data and cache aggregated result
     metas = metaResults.filter(meta => meta && meta.id && meta.name);
     if (metas.length > 0) {
-      await redis?.set(cacheKey, JSON.stringify(metas), { ex: 3600 });
+      const tasks: ListTaskParams[] = [
+        {
+          listName: 'watched',
+          metas,
+          type,
+          traktKey,
+        },
+      ];
+
+      await Promise.all([
+        redis?.set(cacheKey, JSON.stringify(metas), { ex: 3600 }),
+        traktCreateList ? pushListToQstash(tasks) : Promise.resolve(null)
+      ]);
     }
     if (rpdbKey) await updateRpdbPosters(metas, rpdbKey);
     if (backgroundUpdateBatch.length > 0) await pushBatchToQstash(backgroundUpdateBatch);