+++ middleware/googleKeyMiddleware.ts
@@ -125,6 +125,7 @@ export const googleKeyMiddleware = async <P extends Record<string, string | unde
     ctx.state.traktKey = keys.traktKey;
     ctx.state.userId = keys.userId;
     ctx.state.omdbKey = keys.omdbKey || String(OMDB_API_KEY);
+    ctx.state.traktCreateList = keys.traktCreateList;
 
     await next();
   } catch (error) {